refactor: create setup files
