+++ lib/route.js
@@ -53,7 +53,10 @@ const { createChildLogger } = require('./logger-factory.js')
 const { getGenReqId } = require('./reqIdGenFactory.js')
 
 function buildRouting (options) {
-  const router = FindMyWay(options.config)
+  const router = FindMyWay({
+    defaultRoute,
+    ...options.config
+  })
 
   let avvio
   let fourOhFour
@@ -102,6 +105,19 @@ function buildRouting (options) {
     isAsyncConstraint,
     findRoute
   }
+
+  // If the router does not match any route, every request will land here
+    // req and res are Node.js core objects
+    function defaultRoute (req, res) {
+      if (req.headers['accept-version'] !== undefined) {
+        // we remove the accept-version header for performance result
+        // because we do not want to go through the constraint checking
+        // the usage of symbol here to prevent any collision on custom header name
+        req.headers[kRequestAcceptVersion] = req.headers['accept-version']
+        req.headers['accept-version'] = undefined
+      }
+      fourOhFour.router.lookup(req, res)
+    }
 
   function addConstraintStrategy (strategy) {
     throwIfAlreadyStarted('Cannot add constraint strategy!')
